refactor: split turn-resolution session, region geometry, train chrome (#3878)
Decompose GameMapArea next-turn worker session/terminal apply, province
topology segment scan, region-map palette/label placement, production
breakdown DataTable assembly, and train-dialog unit-row widgets into
library parts without changing public behavior.

Refs #3878
